+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 * Convert TODO to https://github.com/michael-cannon/edd-retroactive-licensing/issues
+* Convert to static class coding
 * Revise readme structure
 * Verbiage correction